Made the view more "two-player-friendly".
I commented out the MenuBar at the bottom of the screen, moved the DiceBar and the SettingsButton in the middle and added a 8dip margin at the bottom of the last poison bar. 

Unfortunately I was unable to find a uncomplicated solution to make the DiceBar and the SettingsButton the same size without giving them a fixed height.